adding test for sic mapper
